rename estat-japan to e-stat-japan
