+++ example/e-stat-japan.rb
