+++ lib/datasets.rb
@@ -3,7 +3,7 @@ require_relative "datasets/version"
 require_relative "datasets/adult"
 require_relative "datasets/cifar"
 require_relative "datasets/communities"
-require_relative "datasets/estat-japan"
+require_relative "datasets/e-stat-japan"
 require_relative "datasets/fashion-mnist"
 require_relative "datasets/hepatitis"
 require_relative "datasets/iris"
