+++ test/test-e-stat-japan.rb
